+++ server/src/main/java/luan/moonvs/models/exceptions/IllegalIdException.java
@@ -1,0 +1,8 @@
+package luan.moonvs.models.exceptions;
+
+public class IllegalIdException extends Throwable {
+    public String message;
+    public IllegalIdException(String message) {
+        this.message = message;
+    }
+}