+++ server/src/main/java/luan/moonvs/models/requests/UserAccountRequest.java
@@ -1,0 +1,8 @@
+package luan.moonvs.models.requests;
+
+import java.util.UUID;
+
+public record UserAccountRequest(UUID idUser,
+                                 String username,
+                                 String email,
+                                 String password) { }