+++ server/src/main/java/luan/moonvs/models/responses/AccountResponse.java
@@ -16,4 +16,8 @@ public record AccountResponse(String username,
     public AccountResponse(User user, String message) {
         this(user.getUsername(), user.getEmail(), message);
     }
+
+    public AccountResponse(UserAccountResponse userAccountResponse) {
+        this(userAccountResponse.username(), userAccountResponse.email(), userAccountResponse.message());
+    }
 }