+++ server/src/main/java/luan/moonvs/services/AccountService.java
@@ -5,6 +5,7 @@ import luan.moonvs.models.entities.User;
 import luan.moonvs.models.requests.AccountRequest;
 import luan.moonvs.models.requests.AuthRequest;
 import luan.moonvs.models.responses.AccountResponse;
+import luan.moonvs.models.responses.UserAccountResponse;
 import luan.moonvs.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,23 +19,35 @@ import org.springframework.stereotype.Service;
 @Service
 public class AccountService {
     private final UserRepository repository;
+    private final String SUCCESS_MESSAGE = "User updated successfully";
+
+    @Deprecated
     private final UserBuilder builder;
 
+    @Deprecated
     @Autowired
     private AccountService(UserRepository repository, UserBuilder builder) {
         this.repository = repository;
         this.builder = builder;
     }
 
+    @Deprecated
     public ResponseEntity<AccountResponse> account() {
-        final User authUser = getAuthenticatedUser();
+        final User authUser = getAccount();
         AccountResponse account = new AccountResponse(authUser);
         return ResponseEntity.ok(account);
     }
 
+    public User getAccount() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails userDetails = (UserDetails) auth.getPrincipal();
+        return repository.getUserByUsername(userDetails.getUsername());
+    }
+
+    @Deprecated
     public ResponseEntity<AccountResponse> updateAccount(AccountRequest accountRequest) {
         final String SUCCESS = "Usuário alterado com sucesso!";
-        final User authUser = getAuthenticatedUser();
+        final User authUser = getAccount();
 
         try {
             User user = builder
@@ -54,10 +67,25 @@ public class AccountService {
         }
     }
 
+    public UserAccountResponse updateUsernameAndOrEmail(String username, String email, User user) {
+        try {
+            user = UserBuilder.create(repository, user)
+                    .withUsername(username)
+                    .withEmail(email)
+                    .build();
+
+            repository.save(user);
+            return new UserAccountResponse(HttpStatus.OK, user, SUCCESS_MESSAGE);
+        } catch (IllegalArgumentException e) {
+            return new UserAccountResponse(HttpStatus.BAD_REQUEST, e.getMessage());
+        }
+    }
+
+    @Deprecated
     public ResponseEntity<String> updatePassword(String password) {
         final String SUCCESS = "Senha atualizada com sucesso!",
                 EMPTY_PASSWORD = "Digite uma senha para continuar!";
-        final User authUser = getAuthenticatedUser();
+        final User authUser = getAccount();
 
         if (password == null || password.isBlank())
             return ResponseEntity
@@ -81,9 +109,23 @@ public class AccountService {
         }
     }
 
+    public UserAccountResponse updatePassword(String password, User user) {
+        try {
+            user = UserBuilder.create(repository, user)
+                    .withPassword(password)
+                    .build();
+
+            repository.save(user);
+            return new UserAccountResponse(HttpStatus.OK, SUCCESS_MESSAGE);
+        } catch (IllegalArgumentException e) {
+            return new UserAccountResponse(HttpStatus.BAD_REQUEST, e.getMessage());
+        }
+    }
+
+    @Deprecated
     public ResponseEntity<String> deleteAccount(AuthRequest authRequest) {
         final String EMPTY_STRING = "Forneça %s para prosseguir com a solicitação!";
-        User authUser = getAuthenticatedUser();
+        User authUser = getAccount();
         if (authRequest.username() == null || authRequest.username().isBlank())
             return ResponseEntity
                     .status(HttpStatus.BAD_REQUEST)
@@ -108,9 +150,17 @@ public class AccountService {
         return ResponseEntity.status(HttpStatus.OK).body("Conta deletada com sucesso!");
     }
 
-    protected User getAuthenticatedUser() {
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        UserDetails userDetails = (UserDetails) auth.getPrincipal();
-        return repository.getUserByUsername(userDetails.getUsername());
+    public UserAccountResponse deleteAccount(String username, String password, User user) {
+        final String INVALID_FIELD = "The filled %s doesn't match!",
+                     DELETE_SUCCESS = "The account was successfully deleted!";
+
+        if (!username.equals(user.getUsername()))
+            return new UserAccountResponse(HttpStatus.BAD_REQUEST, String.format(INVALID_FIELD, "username"));
+
+        if (!(new BCryptPasswordEncoder().matches(password, user.getPassword())))
+            return new UserAccountResponse(HttpStatus.BAD_REQUEST, String.format(INVALID_FIELD, "password"));
+
+        repository.delete(user);
+        return new UserAccountResponse(HttpStatus.OK, DELETE_SUCCESS);
     }
 }